fix: refactor itemscripts to safe functions, fix missing returns

+++ battle_royal/battle_royal.h
@@ -89,7 +89,7 @@ private:
     bool gameStarted = false;
     bool aggressivePlayers = false;
     bool startInfecting = false;
-    ulong playerLimit = 40;
+    int playerLimit = 40;
     int currentIteration = 1;
     int startTimer = 60000;
     int relicSwitchTimer = 60000;
